requirements.txt
celery>=3.0.23
requests==2.5.0
python-magic==0.4.6
pushbullet.py==0.5.0
